requirements.txt
Django==3.2.9
captcha==0.4

django-daterange-filter==1.3.0
djangorestframework==3.13.1
django-extra-fields==3.0.2
django-filter==21.1
django-imagekit==4.1.0
django-simple-captcha==0.5.17

drf-yasg==1.20.0
python-dateutil==2.8.2
psycopg2==2.9.3